// PetTemplate.Web/web/src/app/shared/enums/permission.enum.ts
export enum PermissionEnum {
  None = 0,
  PermissionManagement = 1,
  permissionManagement = 1,
  organizationCrud = 2,
  issueRead = 3,
  issueCrud = 4,
  epicRead = 5,
  epicCrud = 6,
  sprintRead = 7,
  sprintCrud = 8,
  hourlyPayRead = 9,
  hourlyPayCrud = 10,
  closePeriodRead = 11,
  closePeriodCrud = 12,
  finPeriodRead = 13,
  finPeriodEdit = 14,
  organizationMemberRead = 15,
  organizationMemberCrud = 16,
  organizationCurrencyRead = 17,
  organizationCurrencyCrud = 18,
}